+++ libs/collapsible/src/lib/collapsible/collapsible.component.ts
@@ -5,6 +5,8 @@ import {
   Output,
   ViewEncapsulation,
   EventEmitter,
+  ContentChild,
+  TemplateRef,
 } from '@angular/core';
 import { ToBoolean } from '@spryker/utils';
 
@@ -20,8 +22,11 @@ export class CollapsibleComponent {
   @Input() titleIcon = '';
   @Input() @ToBoolean() active = false;
   @Input() @ToBoolean() disabled = false;
+  @Input() @ToBoolean() alwaysRender = true;
 
   @Output() activeChange = new EventEmitter<boolean>();
+
+  @ContentChild(TemplateRef) contentTemplate?: TemplateRef<any>;
 
   collapse(): void {
     this.active = false;