+++ tools/scripts/lerna-util.js
@@ -1,1 +1,11 @@
-module.exports.getPackageName = () => require('path').basename(process.cwd());
+module.exports.getPackageName = (scope = 'spryker') => {
+  const packageName = process.env.LERNA_PACKAGE_NAME;
+
+  if (!packageName) {
+    throw new Error(
+      `Package name does not exist in variable 'LERNA_PACKAGE_NAME'!`,
+    );
+  }
+
+  return packageName.replace(`@${scope}/`, '');
+};