+++ libs/collapsible/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-next.2](https://github.com/spryker/ui-components/compare/@spryker/collapsible@0.2.0-next.1...@spryker/collapsible@0.2.0-next.2) (2020-07-29)
+
+**Note:** Version bump only for package @spryker/collapsible
+
+
+
+
+
+# [0.2.0-next.1](https://github.com/spryker/ui-components/compare/@spryker/collapsible@0.2.0-next.0...@spryker/collapsible@0.2.0-next.1) (2020-07-29)
+
+**Note:** Version bump only for package @spryker/collapsible
+
+
+
+
+
 # [0.2.0-next.0](https://github.com/spryker/ui-components/compare/@spryker/collapsible@0.2.0-fix-web-components.0...@spryker/collapsible@0.2.0-next.0) (2020-07-21)